+++ docs/history.rst
@@ -1,4 +1,8 @@
 Version History
 ---------------
+- 0.2.0 [2014-10-30]
+ - Add application level progress bar
+ - Add process start timer reset
+ - Documentation updates
 - 0.1.0 [2014-10-30]
  - Initial version